requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
discord.py==1.7.3
et-xmlfile==1.1.0
idna==3.2
multidict==5.1.0
numpy==1.21.2
openai==0.10.3
openpyxl==3.0.7
pandas==1.3.2
pandas-stubs==1.2.0.11
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
tqdm==4.62.2
typing-extensions==3.10.0.2
urllib3==1.26.6
yarl==1.6.3
